re_tuid
Struct Tuid

#[repr(C, align(1))]
pub struct Tuid {
    pub(crate) time_nanos: [u8; 8],
    pub(crate) inc: [u8; 8],
}

TUID: Time-based Unique Identifier.

Time-ordered globally unique 128-bit identifiers.

The raw bytes of the Tuid sorts in time order as the Tuid itself, and the Tuid is byte-aligned so you can just transmute between Tuid and raw bytes.

Fields
time_nanos: [u8; 8]

Approximate nanoseconds since epoch.

A big-endian u64 encoded as bytes to keep the alignment of Tuid to 1.

We use big-endian so that the raw bytes of the Tuid sorts in time order.

inc: [u8; 8]

Initialized to something random on each thread, then incremented for each new Tuid being allocated.

Uses big-endian u64 encoded as bytes to keep the alignment of Tuid to 1.

We use big-endian so that the raw bytes of the Tuid sorts in creation order.

Implementations

impl Tuid

pub const ARROW_EXTENSION_NAME: &'static str = "rerun.datatypes.TUID"

We give an actual name to Tuid, and inject that name into the Arrow datatype extensions, as a hack so that we can compactly format them when printing Arrow data to the terminal. Check out re_format_arrow for context.

impl Tuid

pub const ZERO: Self

All zeroes.

pub const MAX: Self

All ones.

pub fn new() -> Self

Create a new unique Tuid based on the current time.

pub fn from_nanos_and_inc(time_nanos: u64, inc: u64) -> Self

Construct a Tuid from the upper and lower halves of a u128-bit. The first should be nano-seconds since epoch.

pub fn from_u128(id: u128) -> Self

pub fn as_u128(&self) -> u128

pub fn from_bytes(bytes: [u8; 16]) -> Self

pub fn as_bytes(&self) -> [u8; 16]

Returns most significant byte first (big endian).

pub fn nanos_since_epoch(&self) -> u64

Approximate nanoseconds since unix epoch.

The upper 64 bits of the Tuid.

pub fn inc(&self) -> u64

The increment part of the Tuid.

The lower 64 bits of the Tuid.

pub fn next(&self) -> Self

Returns the next logical Tuid.

Wraps the monotonically increasing back to zero on overflow.

Beware: wrong usage can easily lead to conflicts. Prefer Tuid::new when unsure.

pub fn incremented_by(&self, n: u64) -> Self

Returns the n-next logical Tuid.

This is equivalent to calling Tuid::next n times. Wraps the monotonically increasing back to zero on overflow.

Beware: wrong usage can easily lead to conflicts. Prefer Tuid::new when unsure.

pub fn short_string(&self) -> String

A shortened string representation of the Tuid.